Move PushConfig to the `appconfig` module.
